requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-ckeditor==6.7.3
django-environ==0.12.0
django-js-asset==3.1.2
idna==3.11
pillow==11.3.0
psycopg2==2.9.11
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
